requirements.txt
# Google Agent Developer Kit
google-adk>=1.0.0

# Model Context Protocol (MCP)
mcp>=1.0.0
model-context-protocol>=1.0.0

# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
aiofiles>=23.0.0
aiohttp>=3.8.0
websockets>=11.0.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine Learning
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0

# Materials Science
pymatgen>=2023.5.10
ase>=3.22.0
mattersim>=1.0.0

# Database and Storage
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0

# Web scraping and APIs
beautifulsoup4>=4.12.0
requests>=2.31.0
arxiv>=1.4.0
scholarly>=1.7.0

# Document processing
pypdf>=3.12.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Async and concurrency
asyncio-mqtt>=0.13.0
celery>=5.3.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx-test>=0.1.0

# Development tools
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Configuration
pyyaml>=6.0
toml>=0.10.2

# Visualization and reporting
jinja2>=3.1.0
weasyprint>=59.0
markdown>=3.4.0

# Crystal structure tools (placeholder for CrystaLLM integration)
# crystallm  # Will be installed separately

# Optional: Jupyter for development
jupyter>=1.0.0
ipykernel>=6.25.0
